+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/validation/model/Validation.kt
@@ -1,4 +1,4 @@
-package com.r0adkll.deckbuilder.arch.domain.features.decks.model
+package com.r0adkll.deckbuilder.arch.domain.features.validation.model
 
 
 import paperparcel.PaperParcel
@@ -8,7 +8,8 @@ import paperparcel.PaperParcelable
 @PaperParcel
 data class Validation(
         val standard: Boolean,
-        val expanded: Boolean
+        val expanded: Boolean,
+        val rules: List<Int>
 ) : PaperParcelable {
     companion object {
         @JvmField val CREATOR = PaperParcelValidation.CREATOR