+++ app/src/main/java/com/r0adkll/deckbuilder/arch/domain/features/validation/repository/DeckValidator.kt
@@ -1,0 +1,11 @@
+package com.r0adkll.deckbuilder.arch.domain.features.validation.repository
+
+import com.r0adkll.deckbuilder.arch.domain.features.cards.model.PokemonCard
+import com.r0adkll.deckbuilder.arch.domain.features.validation.model.Validation
+import io.reactivex.Observable
+
+
+interface DeckValidator {
+
+    fun validate(cards: List<PokemonCard>): Observable<Validation>
+}