+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/carddetail/CardDetailRenderer.kt
@@ -2,6 +2,7 @@ package com.r0adkll.deckbuilder.arch.ui.features.carddetail
 
 
 import com.r0adkll.deckbuilder.arch.ui.components.renderers.DisposableStateRenderer
+import com.r0adkll.deckbuilder.util.extensions.mapNullable
 import com.r0adkll.deckbuilder.util.extensions.plusAssign
 import io.reactivex.Scheduler
 
@@ -13,6 +14,22 @@ class CardDetailRenderer(
 ) : DisposableStateRenderer<CardDetailUi.State>(main, comp) {
 
     override fun start() {
+
+        disposables += state
+                .mapNullable { s ->
+                    s.deck?.filter { it.id == s.card?.id }
+                }
+                .distinctUntilChanged()
+                .addToLifecycle()
+                .subscribe { cards ->
+                    actions.showCopies(cards.value?.size)
+                }
+
+        disposables += state
+                .mapNullable { it.deck }
+                .distinctUntilChanged()
+                .addToLifecycle()
+                .subscribe { actions.setEditResults(it.value) }
 
         disposables += state
                 .map { it.validation }