+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/carddetail/di/CardDetailModule.kt
@@ -17,6 +17,10 @@ class CardDetailModule(val activity: CardDetailActivity) {
 
 
     @Provides @ActivityScope
+    fun provideIntentions(): CardDetailUi.Intentions = activity
+
+
+    @Provides @ActivityScope
     fun provideActions(): CardDetailUi.Actions = activity