+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/unifiedsearch/SearchPresenter.kt
@@ -8,6 +8,7 @@ import com.r0adkll.deckbuilder.arch.ui.components.presenter.Presenter
 import com.r0adkll.deckbuilder.arch.ui.features.unifiedsearch.SearchUi.State
 import com.r0adkll.deckbuilder.arch.ui.features.unifiedsearch.SearchUi.State.Change
 import com.r0adkll.deckbuilder.internal.di.scopes.FragmentScope
+import com.r0adkll.deckbuilder.util.extensions.logState
 import com.r0adkll.deckbuilder.util.extensions.plusAssign
 import io.pokemontcg.model.SuperType
 import io.reactivex.Observable
@@ -38,7 +39,7 @@ class SearchPresenter @Inject constructor(
                 .doOnNext { Timber.d(it.logText) }
 
         disposables += merged.scan(ui.state, State::reduce)
-                .doOnNext { state -> Timber.v("    --- $state") }
+                .logState()
                 .subscribe(ui::render)
     }