+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.1.0
+* Improved the builder interface to show the non-evolution line Pokémons in a grid pattern similar to trainer and energy cards
+* Improved UI in the card detail screen to allow you to add more copies of that card, or of variants and evolved forms
+* Removed hard limits on deck rules in favor of soft warnings.
+* Added basic Standard/Expanded format indicators in deck builder interface
+
+# 1.0.0
+* Fixed Quickstart view on tablets
+* Fixed search when just typing 'n' or 'N' to wrap in quotes
+* Fixed issue where '-' and '+' buttons weren't showing in search results
+* Misc. Bugs and copy changes
+
 # 0.2.3
 * Fixed decklist importer to find missing basic energy
 * Fixed keyboard behavior on the deck builder screen